pyproject.toml
[project]
name = "chronicler-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strawberry-graphql[debug-server,fastapi]>=0.257.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["chronicler_backend"]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "flake8>=7.1.1",
    "isort>=5.13.2",
    "pre-commit>=4.0.1",
    "pylint>=3.3.3",
]
test = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
]

[tool.black]
line-length = 100
extend-exclude = "notebooks"

[tool.isort]
skip = ["__init__.py", ".venv"]
profile = "black"
